Refactor DeviceRedirectionMITM's helper classes
