+++ pyrdp/mitm/DeviceRedirectionMITM.py
@@ -4,145 +4,29 @@
 # Licensed under the GPLv3 or later.
 #
 
-import datetime
 import hashlib
 import json
 from logging import LoggerAdapter
 from pathlib import Path
-from typing import BinaryIO, Dict
+from typing import Dict
 
-import names
-
-from pyrdp.core import decodeUTF16LE
+from pyrdp.core import decodeUTF16LE, FileProxy
 from pyrdp.enum import CreateOption, DeviceType, FileAccess, IOOperationSeverity
 from pyrdp.layer import DeviceRedirectionLayer
 from pyrdp.mitm.config import MITMConfig
+from pyrdp.mitm.FileMapping import FileMapping, FileMappingDecoder, FileMappingEncoder
 from pyrdp.parser import DeviceRedirectionParser
-from pyrdp.pdu import DeviceCloseRequestPDU, DeviceCreateRequestPDU, \
-    DeviceIORequestPDU, DeviceIOResponsePDU, DeviceListAnnounceRequest, DeviceReadRequestPDU, DeviceRedirectionPDU
-
-
-class FileMapping:
-    """
-    Class containing information for a file intercepted by the DeviceRedirectionMITM.
-    """
-
-    def __init__(self, remotePath: Path, localPath: Path, creationTime: datetime.datetime, fileHash: str):
-        """
-        :param remotePath: the path of the file on the original machine
-        :param localPath: the path of the file on the intercepting machine
-        :param creationTime: the creation time of the local file
-        :param fileHash: the file hash in hex format (empty string if the file is not complete)
-        """
-        self.remotePath = remotePath
-        self.localPath = localPath
-        self.creationTime = creationTime
-        self.hash: str = fileHash
-
-    @staticmethod
-    def generate(remotePath: Path, outDir: Path):
-        localName = f"{names.get_first_name()}{names.get_last_name()}"
-        creationTime = datetime.datetime.now()
-
-        index = 2
-        suffix = ""
-
-        while True:
-            if not (outDir / f"{localName}{suffix}").exists():
-                break
-            else:
-                suffix = f"_{index}"
-                index += 1
-
-        localName += suffix
-
-        return FileMapping(remotePath, outDir / localName, creationTime, "")
-
-
-class FileMappingEncoder(json.JSONEncoder):
-    """
-    JSON encoder for FileMapping objects.
-    """
-
-    def default(self, o):
-        if isinstance(o, datetime.datetime):
-            return o.isoformat()
-        elif not isinstance(o, FileMapping):
-            return super().default(o)
-
-        return {
-            "remotePath": str(o.remotePath),
-            "localPath": str(o.localPath),
-            "creationTime": o.creationTime,
-            "sha1": o.hash
-        }
-
-
-class FileMappingDecoder(json.JSONDecoder):
-    """
-    JSON decoder for FileMapping objects.
-    """
-
-    def __init__(self):
-        super().__init__(object_hook=self.decodeFileMapping)
-
-    def decodeFileMapping(self, dct: Dict):
-        for key in ["remotePath", "localPath", "creationTime"]:
-            if key not in dct:
-                return dct
-
-        creationTime = datetime.datetime.strptime(dct["creationTime"], "%Y-%m-%dT%H:%M:%S.%f")
-        return FileMapping(Path(dct["remotePath"]), Path(dct["localPath"]), creationTime, dct["sha1"])
-
-
-class FileProxy:
-    """
-    Proxy object that waits until a file is accessed before creating it.
-    """
-
-    def __init__(self, path: Path, mode: str, mapping: FileMapping, log: LoggerAdapter):
-        """
-        :param path: path of the file
-        :param mode: file opening mode
-        :param mapping: FileMapping object for this file
-        :param log: logger for this component
-        """
-        self.path = path
-        self.mode = mode
-        self.mapping = mapping
-        self.log = log
-        self.file: BinaryIO = None
-
-    def createFile(self):
-        """
-        Create the file and overwrite this object's methods with the file object's methods.
-        """
-
-        if self.file is None:
-            self.log.info("Saving file '%(remotePath)s' to '%(localPath)s'", {"localPath": self.path, "remotePath": self.mapping.remotePath})
-
-            self.file = open(str(self.path), self.mode)
-            self.write = self.file.write
-            self.seek = self.file.seek
-            self.close = self.file.close
-
-    def write(self, *args, **kwargs):
-        self.createFile()
-        self.file.write(*args, **kwargs)
-
-    def seek(self, *args, **kwargs):
-        self.createFile()
-        self.file.seek(*args, **kwargs)
-
-    def close(self):
-        if self.file is not None:
-            self.log.debug("Closing file %(path)s", {"path": self.path})
-            self.file.close()
+from pyrdp.pdu import DeviceCloseRequestPDU, DeviceCreateRequestPDU, DeviceIORequestPDU, \
+    DeviceIOResponsePDU, DeviceListAnnounceRequest, DeviceReadRequestPDU, DeviceRedirectionPDU
 
 
 class DeviceRedirectionMITM:
     """
     MITM component for the device redirection channel.
+    It saves files transferred over RDP to a local directory. The files aren't named after their remote name to avoid
+    conflicts. Rather, they are given a random name, and the mapping to their remote path is given by the mapping.json
+    file. Each unique file (identified by its hash) is saved only once. Duplicates are removed to avoid filling the drive
+    with identical files.
     """
 
     def __init__(self, client: DeviceRedirectionLayer, server: DeviceRedirectionLayer, log: LoggerAdapter, config: MITMConfig):
@@ -152,6 +36,7 @@ class DeviceRedirectionMITM:
         :param log: logger for this component
         :param config: MITM configuration
         """
+        super().__init__()
 
         self.client = client
         self.server = server
@@ -159,15 +44,9 @@ class DeviceRedirectionMITM:
         self.config = config
         self.currentIORequests: Dict[int, DeviceIORequestPDU] = {}
         self.openedFiles: Dict[int, FileProxy] = {}
+        self.openedMappings: Dict[int, FileMapping] = {}
         self.fileMap: Dict[str, FileMapping] = {}
-
-        try:
-            with open(self.filemapFile, "r") as f:
-                self.fileMap: Dict[str, FileMapping] = json.loads(f.read(), cls=FileMappingDecoder)
-        except IOError:
-            pass
-        except json.JSONDecodeError:
-            self.log.error(f"Failed to decode file mapping, overwriting previous file")
+        self.fileMapPath = self.config.outDir / "mapping.json"
 
         self.client.createObserver(
             onPDUReceived=self.onClientPDUReceived,
@@ -177,19 +56,20 @@ class DeviceRedirectionMITM:
             onPDUReceived=self.onServerPDUReceived,
         )
 
-    @property
-    def filemapFile(self) -> str:
-        """
-        Get the path to the saved file mapping.
-        """
-        return str(self.config.outDir / "mapping.json")
+        try:
+            with open(self.fileMapPath, "r") as f:
+                self.fileMap: Dict[str, FileMapping] = json.loads(f.read(), cls=FileMappingDecoder)
+        except IOError:
+            pass
+        except json.JSONDecodeError:
+            self.log.error(f"Failed to decode file mapping, overwriting previous file")
 
     def saveMapping(self):
         """
         Save the file mapping to a file in JSON format.
         """
 
-        with open(self.filemapFile, "w") as f:
+        with open(self.fileMapPath, "w") as f:
             f.write(json.dumps(self.fileMap, cls=FileMappingEncoder, indent=4, sort_keys=True))
 
     def onClientPDUReceived(self, pdu: DeviceRedirectionPDU):
@@ -214,19 +94,6 @@ class DeviceRedirectionMITM:
 
         destination.sendPDU(pdu)
 
-    def handleDeviceListAnnounceRequest(self, pdu: DeviceListAnnounceRequest):
-        """
-        Log mapped devices.
-        :param pdu: the device list announce request.
-        """
-
-        for device in pdu.deviceList:
-            self.log.info("%(deviceName)s mapped with ID %(deviceID)d: %(deviceData)s", {
-                "deviceName": DeviceType.getPrettyName(device.deviceType),
-                "deviceID": device.deviceID,
-                "deviceData": device.preferredDosName.rstrip(b"\x00").decode()
-            })
-
     def handleIORequest(self, pdu: DeviceIORequestPDU):
         """
         Keep track of IO requests that are active.
@@ -234,6 +101,7 @@ class DeviceRedirectionMITM:
         """
 
         self.currentIORequests[pdu.completionID] = pdu
+
 
     def handleIOResponse(self, pdu: DeviceIOResponsePDU):
         """
@@ -257,6 +125,19 @@ class DeviceRedirectionMITM:
         else:
             self.log.error("Received IO response to unknown request #%(completionID)d", {"completionID": pdu.completionID})
 
+    def handleDeviceListAnnounceRequest(self, pdu: DeviceListAnnounceRequest):
+        """
+        Log mapped devices.
+        :param pdu: the device list announce request.
+        """
+
+        for device in pdu.deviceList:
+            self.log.info("%(deviceName)s mapped with ID %(deviceID)d: %(deviceData)s", {
+                "deviceName": DeviceType.getPrettyName(device.deviceType),
+                "deviceID": device.deviceID,
+                "deviceData": device.preferredDosName.rstrip(b"\x00").decode()
+            })
+
     def handleCreateResponse(self, request: DeviceCreateRequestPDU, response: DeviceIOResponsePDU):
         """
         Prepare to intercept a file: create a FileProxy object, which will only create the file when we actually write
@@ -274,9 +155,17 @@ class DeviceRedirectionMITM:
         if isFileRead and isNotDirectory:
             remotePath = Path(decodeUTF16LE(request.path).rstrip("\x00"))
             mapping = FileMapping.generate(remotePath, self.config.fileDir)
+            proxy = FileProxy(mapping.localPath, "wb")
 
-            localPath = mapping.localPath
-            self.openedFiles[response.fileID] = FileProxy(localPath, "wb", mapping, self.log)
+            self.openedFiles[response.fileID] = proxy
+            self.openedMappings[response.fileID] = mapping
+
+            proxy.createObserver(
+                onFileCreated = lambda _: self.log.info("Saving file '%(remotePath)s' to '%(localPath)s'", {
+                    "localPath": mapping.localPath, "remotePath": mapping.remotePath
+                }),
+                onFileClosed = lambda _: self.log.debug("Closing file %(path)s", {"path": mapping.localPath})
+            )
 
 
     def handleReadResponse(self, request: DeviceReadRequestPDU, response: DeviceIOResponsePDU):
@@ -292,8 +181,13 @@ class DeviceRedirectionMITM:
             file.seek(request.offset)
             file.write(response.readData)
 
-            self.fileMap[file.mapping.localPath.name] = file.mapping
-            self.saveMapping()
+            # Save the mapping permanently
+            mapping = self.openedMappings[request.fileID]
+            fileName = mapping.localPath.name
+
+            if fileName not in self.fileMap:
+                self.fileMap[fileName] = mapping
+                self.saveMapping()
 
     def handleCloseResponse(self, request: DeviceCloseRequestPDU, _: DeviceIOResponsePDU):
         """
@@ -310,8 +204,10 @@ class DeviceRedirectionMITM:
             if file.file is None:
                 return
 
+            currentMapping = self.openedMappings.pop(request.fileID)
+
             # Compute the hash for the final file
-            with open(str(file.mapping.localPath), "rb") as f:
+            with open(currentMapping.localPath, "rb") as f:
                 sha1 = hashlib.sha1()
 
                 while True:
@@ -322,16 +218,16 @@ class DeviceRedirectionMITM:
 
                     sha1.update(buffer)
 
-                file.mapping.hash = sha1.hexdigest()
+                currentMapping.hash = sha1.hexdigest()
 
             # Check if a file with the same hash exists. If so, keep that one and remove the current file.
             for localPath, mapping in self.fileMap.items():
-                if mapping is file.mapping:
+                if mapping is currentMapping:
                     continue
 
-                if mapping.hash == file.mapping.hash:
-                    file.mapping.localPath.unlink()
-                    self.fileMap.pop(file.mapping.localPath.name)
+                if mapping.hash == currentMapping.hash:
+                    currentMapping.localPath.unlink()
+                    self.fileMap.pop(currentMapping.localPath.name)
                     break
 
             self.saveMapping()
